hashing: Filter-out paths that are not relative to model_path
When there is an existing signature in 'model-dir/model.sig' but the new
signature is to be written into './model.sig', then that old signature
file must be part of the new signature. For this we need to filter-out all
ignore paths, which include the signature, and remove all those that are
not relative to the model-dir.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>