Record and apply ignored paths (#462)
* hashing: Fix the paths to the git related files

Commit 8aba2adada812 made changes so that git-related files are not
ignored anymore. Fix this by fixing their paths.

Resolves: https://github.com/sigstore/model-transparency/issues/453
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* signing: Record ignored paths as part of the predicate in the signature

Record ignored paths as part of the serialization info in the predicate.
It may now look like this:

  "predicateType": "https://model_signing/signature/v1.0",
  "predicate": {
    "serialization": {
      "method": "files",
      "hash_type": "sha256",
      "allow_symlinks": false
      "ignore_paths": [
        "model.sig",
        ".git",
        ".gitattributes",
        ".github",
        ".gitignore"
      ],
    },

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* verifying: Use ignored paths from predicate in the signature

Use the signer's ignored paths from the the predicate's serialization info
in the signature and apply them when ignoring files while hashing.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* hashing: Fix paths of to-be-ignored files by using relpath

To avoid making an existing signature file part of the signature by
providing a path like the following, use relpath on each path given by the
user.

   ../curr-dir/model-dir/model.sig

In this case '../curr-dir' is basically a no-op and the path can be
rewritten (by relpath) like this:

   model-dir/model.sig

With this path the filter function should_ignore will properly ignore
the model.sig file.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* hashing: Filter-out paths that are not relative to model_path

When there is an existing signature in 'model-dir/model.sig' but the new
signature is to be written into './model.sig', then that old signature
file must be part of the new signature. For this we need to filter-out all
ignore paths, which include the signature, and remove all those that are
not relative to the model-dir.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

---------

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>