cli: Adjust model name when signing and verifying in model_path '.' (#452)
When signing or verifying in model_path '.' then the model_path.name
becomes an empty string and cannot be used for the name of the model in
the Manifest. In this case, derive the name of the model from the model
paths' absolute path and use the last directory's name. Use the same
method for the model_path '..'.

Extend a test case with signing while in the model directory and add a
test case when signing while in a sub-directory of the model directory
while using '.. for the model path.

Resolves: https://github.com/sigstore/model-transparency/issues/451

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>