Move and extend python tests (#463)
* hashing: Use Path.resolve() rather than os.path.relpath()

On Windows os.path.relpath() may create problems if the given path is on
a different drive than the current working directory. Therefore, use
the Path's resolve() function to normalize weird paths.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* tests: Move python sign+verify test into api_test.py

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* tests: Have git related files ignored

Extend the test cases with checks whether git related files are properly
ignored. Create a .gitignore file in the model directory and ensure that
is part of the signature when ignore_git_paths = False, and not part of
the signature when ignore_git_paths = True.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* tests: Add checks for the recorded list of ignored paths

Extend the test cases and check that the recorded list of ignore_paths
in the signature is as expected. Add a non-existing file to the list
and have it ignored.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* tests: Check that the proper model name is in the subject name

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* tests: Check that the proper model name is in the subject name (bash)

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

---------

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>