Fix handling paths to certs from `certifi.where()`
If `certifi.where()` returns a single string, we were creating multiple
`pathlib.Path` objects for each character in the string, instead of
wrapping the entire path string in `pathlib.Path`.

This is a type error bug, due to the fact that strings in Python are
sometimes lists and sometimes not.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>