add disk vs memory hasher experiment (#478)
This benchmark generates temporary files of varying sizes, and then
hashes the file using a modified `io.SimpleFileHasher` to measure the
time spent reading file chunks into memory, and the time spent to hash
those chunks separately. The disk hasher reports the sum of both of
those times, while the memory hasher only reports the time spent
hashing.

Ensure this experiment is run in a low memory environment to minimize
the effect of OS disk caching.

Signed-off-by: Spencer Schrock <sschrock@google.com>