add max_workers experiment
This benchmarking experiment evaluates the effect of the max_workers
count on sharded hashing. The worker counts were picked to fall around
vCPU counts for common VMs.

Local experiments with a large model (to ignore the effects of OS disk
caching) showed a diminishing effect beyond 8 workers, indicating the
disk became the bottleneck. By default, we defer to the value selected
by the `concurrent.futures` library (cpu_count + 4, up to a maximum of
32).

We likely don't need to increase our value beyond the max of 32, but we
may want to advocate for users to set the value lower if disk usage is a
concern.

Signed-off-by: Spencer Schrock <sschrock@google.com>