add benchmark trigger workflow
This is modeled after the OpenSSF Scorecard `scdiff` workflow, which
looks for comments from repository members. This requires the developer
triggering the benchmark to have their membership in the Sigstore
organization public.

This approach gains flexibility compared to a label trigger as
additional arguments can be provided after the /bench command.

Signed-off-by: Spencer Schrock <sschrock@google.com>