Add support for ignoring files that are not part of signature/manifest
Add support for command line option --ignore-unsigned-files that allows to
ignore files that are not part of the signature (= not listed in the
manifest). This allows to ignore files that for example were added after a
signature was created and those files' presence would make the signature
verification fail. Another use case for this is the presence of multiple
signatures in the same directory where it is necessary to ignore the
files that are not covered by each signature.

Add support for this option for all verification methods. If this
option is set, then create a list of files_to_hash that is derived from
the list of files in the signature manifest.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>