Add support for ignoring files that are not part of signature/manifest
Add support for command line option --ignore-unsigned-files that allows to
ignore files that are not part of the signature (= not listed in the
manifest). This allows to ignore files that for example were added after a
signature was created and those files' presence would make the signature
verification fail. Another use case for this is the presence of multiple
signatures in the same directory where it is necessary to ignore the
files that are not covered by each signature.

Add support for this option for all verification methods. If this option
is set, then use superset comparison of items in the actual manifest versus
the expected one. Make sure that the items in the actual manifest are NOT
a true subset of the items in the expected manifest because if they were
a subset it would mean that files are missing. They may be equal or be a
superset, with superset meaning that there are files that are supposed to
be ignored.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>