Fix various issues (#500)
* tests: Fix several copy and paste errors in error messages

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* _cli: Remove type=bool from --allow_symlinks definition

With click v8.2.0, When type=bool and is_flag=True are both set, then the
--allow_symlinks option does not work properly, which fails the test case
for example. This behavior seems to be fixed in click v8.2.1. Nevertheless,
remove type=bool from the definition of the option since it doesn't seem
to be necessary.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* ci: Run apt update before apt install

Run apt update before apt install to avoid issues with updated packages.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

---------

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>