Add more tests (#537)
* tests: Add test cases for signatures created by v1.0.1

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* tests: Add testing of signatures across many different library versions

Create a signature with the currently active model signing library and test
it against old versions of the library installed into a venv. Use the older
versions of the library to test against the pre-created signatures located
in the test directory.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

* tests: Add suffix .sh to all bash scripted tests

Also adjust the testrunner to pick up test cases with suffix .sh.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>

---------

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>