Add blake3 hashing (#538)
Because BLAKE3 natively supports parallelism without changing the final
hash, sharding is bypassed. This is much more useful than getting
different file hashes depending on which hashing method you used.

The BLAKE3 hashing is done by memory mapping the file, and defaults to
the max number of workers which is the number of logical CPU cores.
This is a good default and the most performant setup. It is also what
the standard BLAKE3 CLI tool (b3sum) does. It is implemented in Rust
and so will be true parallelism rather than the thread concurrency
implemented for other hashing algorithms, so the speed up should be
quite large. But it will likely be slower on HDDs than having no
parallelism. I think this is the right default, but the HDD concern
is documented.

Resolves: #530

Signed-off-by: makeworld <makeworld@protonmail.com>