fix: remove unintended conditional grouping causing overtriggering
Previously testing if an issue comment was a PR and if the comment
started with `/bench` were grouped in the same expression and evaluated
to true for every PR comment.

The ${{ }} expression syntax is optional in `if` conditionals, so I
removed it.  This wasn't caught initially as I always tested with
comments that started with `/bench`, and never with a comment that
shouldn't have triggered the benchmark.

Signed-off-by: Spencer Schrock <sschrock@google.com>