Remove SLSA for ML for now.
The demo workflows on GitHub Actions are stale and would require
significant work to make work. The GCP ones are not running, so they are
even more stale.

This way, we also focus this repo to be just for model signing. Future
rename to match the package name is possible.

When we start SLSA work again, if needed, we can create new repos, or
resurface some of the code from here.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>